Include Time of Day in Report Step List
When the user passes the '-l' option, the convertECL utility
generates a list of restart steps in a unified restart file.  At
present, this list has the following format:

    Report step number:    0   Date: 1997/11/06
    Report step number:    1   Date: 1997/11/14
    [...]
    Report step number:  239   Date: 2006/09/14
    Report step number:  240   Date: 2006/10/01
    Report step number:  241   Date: 2006/10/10

This commit extends the list to also include the time of day, which
is particularly useful if there are multiple restart steps in a
single day.  We also switch to outputting month names instead of
month numbers similarly to how the rest of the OPM suite reports
dates.  With these changes, the above list will be printed as

    Report step number:    0   Date: 06-Nov-1997 00:00:00
    Report step number:    1   Date: 14-Nov-1997 00:00:00
    [...]
    Report step number:  239   Date: 14-Sep-2006 00:00:00
    Report step number:  240   Date: 01-Oct-2006 00:00:00
    Report step number:  241   Date: 10-Oct-2006 00:00:00

instead.