Revise Summary::Eval() Function Signature
This commit introduces an aggregate for client code to pass dynamic
simulation state objects into the Summary::eval() member function.
The aggregate is a collection of pointers for which nullptr
represents "no such value".  This is a large, but hopefully
one-time, API cost that prepares for introducing additional dynamic
state objects as arguments to Summary::eval().  The short-term use
case is adding a more streamlined container for region level vectors
that in turn will enable removing the COPT-based hack for computing
the ROEW* vectors.

This protocol, while clearly more opaque than the current approach
of having a separate parameter for each object type, is nevertheless
expected to be easier to maintain in the future.