Expose Foam Density Unit
This commit adds a new UnitSystem::measure for "foam density"
quantities such as those entered in the WFOAM keyword.  This is
mainly intended for client code that wishes to tag output arrays
appropriately for unit conversion in restart file output.

While here, also add unit tests for the 'concentration' unit of
measure that was introduced in commit d4e33a461 (PR #4741).