Support well completion quantities in ACTIONX
For this we need to convert e.g. WWCTL__3 on well W1 to WWCTL:W1:3 for
looking up the summary quantity and make sure that the  func_type in
ACTIONX is Well.

In addition we check that number specifier in well completion quantity
has the 3 characters used for the number and throw an expection with
message similar to:

Well completion quantity WWCTL_3 needs to have three consecutive characters that are  a number or an underline character  (e.g. WWCTL123:W1, WWCTL_23:W or WWCTL__3:W1)