Drop Region Set From Vector Names in SummaryState::update() Call
This ensures that the internal logic for generating unique vector
keys always computes the same string value.  In turn, this fixes a
problem concerning inter-region flow vectors with a user defined
region set like

    ROFT_ABC
    ROFT+ABC
    ROFT-ABC

Prior to this commit, these would internally be stored with
duplicated region set names and would be incorrectly written as zero
in the summary files.