Add Support for WORK Arrays in OPERATE and OPERATER
In this initial version we support only the double precision WORK
arrays (named "WORK<n>" with <n> being a positive integer), and only
for compressed property arrays like SOGCR.  In other words,
operating on the "global" MULTZ* arrays is not supported when
referencing WORK arrays at this time.  On the other hand, we do not
use REGDIMS(8) as an upper limit on the number of distinct "WORK"
arrays that can be used in the model description.

Integer work arrays ("IWORK<n>") are not supported at this time.

We add a new data member, FieldProps::work_arrays, of the same type
as FieldProps::double_data, and select this 'work_arrays' member as
the backing store for array references using "WORK<n>".  All WORK
arrays are cleared at the end of processing a section (e.g., GRID,
PROPS, EDIT).